rename test_debugger_base to utils
